feat: 구입 금액 대비 당첨 금액에 따른 수익률을 구한다.

+++ src/main/java/lotto/LottoRank.java
@@ -29,7 +29,7 @@ public enum  LottoRank {
                 .orElse(LottoRank.NONE);
     }
 
-    public int reward() {
+    public long reward() {
         return expression.get();
     }
 }
